+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/FunctionUtils.kt
@@ -9,3 +9,8 @@ val FUNCTIONS_CORE_TOOLS_KNOWN_SUPPORTED_VERSIONS = listOf("v2", "v3", "v4")
 
 // Latest supported version by the Azure Toolkit for Rider
 const val FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION = "v4"
+
+fun isFunctionCoreToolsExecutable(value: String?) =
+    value.equals("func", ignoreCase = true) ||
+            value.equals("func.cmd", ignoreCase = true) ||
+            value.equals("func.exe", ignoreCase = true)