+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/settings/AzureFunctionSettings.kt
@@ -5,39 +5,36 @@
 package com.microsoft.azure.toolkit.intellij.legacy.function.settings
 
 import com.intellij.openapi.components.*
-import java.nio.file.Path
+import com.microsoft.azure.toolkit.intellij.legacy.function.isFunctionCoreToolsExecutable
+import com.microsoft.azure.toolkit.lib.appservice.utils.FunctionCliResolver
 
 @State(
     name = "com.microsoft.azure.toolkit.intellij.legacy.function.settings.AzureFunctionSettings",
     storages = [(Storage("AzureSettings.xml"))]
 )
 @Service
-class AzureFunctionSettings : SimplePersistentStateComponent<AzureFunctionSettingState>(
-    AzureFunctionSettingState(
-        Path.of(System.getProperty("user.home")).resolve(AZURE_TOOLS_FOLDER).resolve(AZURE_FUNCTIONS_TOOLS_FOLDER)
-    )
-) {
+class AzureFunctionSettings : SimplePersistentStateComponent<AzureFunctionSettingState>(AzureFunctionSettingState()) {
     companion object {
         fun getInstance() = service<AzureFunctionSettings>()
 
-        private const val AZURE_TOOLS_FOLDER = ".AzureToolsForIntelliJ"
-        private const val AZURE_FUNCTIONS_TOOLS_FOLDER = "AzureFunctionsCoreTools"
+        const val AZURE_TOOLS_FOLDER = ".AzureToolsForIntelliJ"
+        const val AZURE_FUNCTIONS_TOOLS_FOLDER = "AzureFunctionsCoreTools"
     }
 
     var azureCoreToolsPathEntries: List<AzureCoreToolsPathEntry>
         get() {
             return buildList {
-                add(AzureCoreToolsPathEntry("v4", state.functionV4Path ?: ""))
-                add(AzureCoreToolsPathEntry("v3", state.functionV3Path ?: ""))
-                add(AzureCoreToolsPathEntry("v2", state.functionV2Path ?: ""))
+                add(AzureCoreToolsPathEntry("v4", resolveFromEnvironment(state.functionV4Path)))
+                add(AzureCoreToolsPathEntry("v3", resolveFromEnvironment(state.functionV3Path)))
+                add(AzureCoreToolsPathEntry("v2", resolveFromEnvironment(state.functionV2Path)))
             }
         }
         set(entries) {
             for (entry in entries) {
                 when (entry.functionsVersion) {
-                    "v4" -> state.functionV4Path = entry.coreToolsPath
-                    "v3" -> state.functionV3Path = entry.coreToolsPath
-                    "v2" -> state.functionV2Path = entry.coreToolsPath
+                    "v4" -> state.functionV4Path = resolveFromEnvironment(entry.coreToolsPath)
+                    "v3" -> state.functionV3Path = resolveFromEnvironment(entry.coreToolsPath)
+                    "v2" -> state.functionV2Path = resolveFromEnvironment(entry.coreToolsPath)
                     else -> continue
                 }
             }
@@ -54,6 +51,14 @@ class AzureFunctionSettings : SimplePersistentStateComponent<AzureFunctionSettin
         set(value) {
             state.checkForFunctionMissingPackages = value
         }
+
+    private fun resolveFromEnvironment(coreToolsPathValue: String?): String {
+        if (isFunctionCoreToolsExecutable(coreToolsPathValue)) {
+            return FunctionCliResolver.resolveFunc() ?: ""
+        }
+
+        return coreToolsPathValue ?: ""
+    }
 }
 
 data class AzureCoreToolsPathEntry(var functionsVersion: String, var coreToolsPath: String)