+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/function/templates/InstallFunctionToolComponent.kt
@@ -8,7 +8,6 @@ package com.microsoft.azure.toolkit.intellij.legacy.function.templates
 
 import com.intellij.ide.actions.ShowSettingsUtilImpl
 import com.intellij.openapi.project.ProjectManager
-import com.intellij.openapi.rd.util.withBackgroundContext
 import com.intellij.openapi.util.registry.Registry
 import com.intellij.platform.ide.progress.runWithModalProgressBlocking
 import com.intellij.ui.IdeBorderFactory
@@ -17,6 +16,8 @@ import com.intellij.util.ui.JBInsets
 import com.jetbrains.rider.ui.components.base.Viewable
 import com.microsoft.azure.toolkit.intellij.legacy.function.FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION
 import com.microsoft.azure.toolkit.intellij.legacy.function.coreTools.FunctionCoreToolsInfoProvider
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import javax.swing.JComponent
 import javax.swing.JPanel
 
@@ -37,7 +38,7 @@ class InstallFunctionToolComponent(reloadTemplates: Runnable) : Viewable<JCompon
                 link("Download Azure Functions Core Tools... (recommended)") {
                     val project = ProjectManager.getInstance().defaultProject
                     runWithModalProgressBlocking(project, "Downloading Azure Functions Core Tools...") {
-                        withBackgroundContext {
+                        withContext(Dispatchers.Default) {
                             FunctionCoreToolsInfoProvider.getInstance()
                                 .retrieveForVersion(
                                     FUNCTIONS_CORE_TOOLS_LATEST_SUPPORTED_VERSION,